fix compiler warning

+++ ekrano/src/goldy_renderer.rs
@@ -1523,7 +1523,7 @@ impl GoldyRenderer {
             );
         }
 
-        let mut output = vec![0_u8; texture.byte_size()];
+        let mut output = vec![0_u8; texture.byte_size() as usize];
         texture
             .read_to_cpu(&mut output)
             .map_err(|e| Error::Readback(e.to_string()))?;
